Update CryptoObject usage in Tests
Previously, CryptoObjects were tested pre 35 and 35 using a VIC
emulator, and passed presubmit with no issues. However, that was
utilizing non release such that SDK_INT wasn't 35, even if the emulator
indicated it was running VIC. Thus some test cases needed to be fixed
and the way the crypto object was generated in tests needed to be
changed. This uses a real keystore behind the crypto object, so it is
expected that emulated tests have screen locks. That would match
expected use cases for the Biometric Flow, which is our intended
approach. However, the emulators that run these tests for presubmit do
not contain screenlocks, so a version of the CryptoObject was created
that uses a real Cipher but no longer requires any form of
authentication existing on the device.

Bug: 356875421
Bug: 353536734
Test: These are Unit Tests
Change-Id: I9cc49cb3111706595911339528bb528a2699e374