K2 fixes for collection
The build still fails despite these changes due to workaround from KT-29963 no
longer working.

Note: On annotations annotated with @OptionalExpectation such as @JvmOverloads; the way they work is that they are meant to be used in commonMain, but are allowed to have missing actual implementations, which is how the compiler knows to skip their effect for unsupported platforms. For example, @JvmOverloads only has an effect for JVM. For expect / actuals, k2 compiler throws a suppressible error if you try to use it in native actuals since it is missing. However, we also have a conflicting lint which throws a warning about mismatched annotations as you are also meant to copy all common annotations to their actualized counterparts. As a result the only correct usage of @OptionalExpectation annotation is to use them only on the implementations they apply to. For example, on a fully commonized implementation, or for expect / actual, only on the actuals they apply to.

Note: Replacing the deprecated cleaner api also requires bumping to at least
kotlin 1.9

Bug: 358688708
Test: ./gradlew collection:collection:build -Pandroidx.forceKotlin20Target=true
Change-Id: I67ff4141cdbd8f9a2eade58598b45e03b672f4d5