Introduce a GraphLoop that will accept and submit requests.
This class is built to asyncronously enqueue and process commands
that will be sent to an active camera. This replicates a lot of
the core logic that currently exists in GraphProcessor while also
addressing several race conditions and bugs that can show up with
the existing implementation when requests are aborted or during
fast sequences of events and updates.

The new system is more testable and deterministic, and will likely
reduce the need for additional synchronization in downstream
classes. This will reduce CPU load and overhead when submitting
requests.

Bug: 263211462
Bug: 247166179

Test: ./gradlew\
      :camera:camera-camera2-pipe:testDebugUnitTest\
      :camera:camera-camera2-pipe-testing:testDebugUnitTest\
      :camera:camera-camera2-pipe-integration:testDebugUnitTest
Change-Id: I8ccc1d45820ac217a0e196e90a4bda2ee120cba0