Correctly fixup linked list nodes after tombstone cleanup
When a SieveCache contained enough tombstones to require
dropDeletes() to run and if the entries have hash collisions,
the existing code could corrupt its internal mapping of
older indices to new indices post rehash.

Instead of tracking a simple mapping of previous/new index
to be able to fixup the nodes linked list, this change
tracks a more complex mapping consisting of a pair of
(source, destination) indices per entry. When entries are
moved around the arrays, we use the source mapping to go
update the original mapping to properly track where the
value ends up at the end of the rehash.

See comments in dropDeletes() for more details.

The new behavior is validated by a new unit test that
triggers numerous hash collisions. The test fails without
this CL.

This change also fix the way values were swapped in the arrays
for all "scatter" collections to avoid "leaking" keys.

Relnote: N/A
Fixes: 356648137
Test: SieveCacheTest, androidx.collection tests
Change-Id: I2e54eac22bd41e990d267fa2b3d18339c3b2b49e