Add OpenOnPhoneDialog for Material3.
This adds a new component OpenOnPhoneDialog, which should be used when some action should be continued on the phone.

OpenOnPhoneDialog will show a progress and will be automatically dismissed after a specified timeout.
Bug: 352532262
Test: OpenOnPhoneDialogTest.kt, OpenOnPhoneDialogScreenshotTest.kt
Relnote: "New OpenOnPhoneDialog component is now available for the Wear Material3 library. This component should be used to indicate an action that will continue on the user's phone and will be dismissed after a specified timeout."

Change-Id: I978fdd711362f808cac89da4c15fce337ed100bc