Ink Strokes API: Unbreak build
Fix lint errors that weren't caught during presubmit. This does not have a
corresponding internal CL yet, the fix is being applied directly to
androidx-main to unbreak the build as soon as possible, and an internal CL
will follow.

Test: ./gradlew :ink:ink-brush:lintAndroidMain

Relnote: Update annotations and tests for Brush out of bounds inputs.
Change-Id: Iee1c0532ae6e790045fec7942742065a7071847e