Refactor navigation runtime lint
Copied helper functions from compose lint util to our navigation lint
common to make it easier to create stubs in the future.

Refactored stubs to separate the actual library stubs from the test code
we need to stub to ensure we have correct compilation. This included
some clean up on the test code so that it didn't all exists in the
navigation package.

Removed the test case for only source code. We should always be using
the bytecode for our library stubs. Our test code stubs source is only
beneficial if the source is out of sync with the bytecode, and the
helper functions notifies of any changes to the source code that the
bytecode should be updated.

Since we are now only testing with bytecode, we can remove the
paramaterized tests and simplify to regular junit tests.

Test: all existsing tests have been updated and pass.
Change-Id: I8f816339b74cd4b117efd6c166b4456def66dd46