Ink: Expose Brush v1 APIs
Remove the RestrictTo(LIBRARY_GROUP) annotations from the Brush v1 APIs for
API+2 review. Move those annotations onto experimental APIs for now, as
experimental APIs will still trigger API+2 review and it will be easier if
those are reviewed in a separate CL.

The APIs here reflect the combination of the approved API design document for
brushes and the approved design document for the relationship between Ink Brush
color and Jetpack Compose Color. Any proposed deviations to this API may need to
be discussed with prior approvers of these documents.

This includes a few small updates to the API to satisfy API lint. These will
be backported to the internal code that this is migrating from to keep them
consistent.
- Brush.colorInt is renamed to Brush.colorIntArgb to match the naming of
  Brush.Builder.setColorIntArgb and Brush.copyWithColorIntArgb.
- Reorder params in Brush.copyWithColorLong and Brush.copyWithColorIntArgb
  to put required parameters first.

This also includes small kdoc changes to match the design document that hadn't
made it to the code yet.

Relnote: Make Ink Brush APIs public.

Test: ./gradlew :ink:ink-brush:jvmTest

Change-Id: Ib40f0cb75fcf64fc411ceb6079b8ee184a9a72fc