Moving raw query methods to use drivers.
The aim of this CL is to slowly deprecate reliance on
the `useDriverApi` flag, starting with the Android-only raw query methods that receive SupportSQLiteQuery.

This was done by migrating these raw queries to use driver API's by converting them to RoomRawQuery.

Bug: 319660042
Test: DaoKotlinCodeGenTest.kt, existing instrumentation tests.
Relnote: Added code gen utility function to convert from SupportSQLite to RoomRawQuery.
Change-Id: I4292a45b0e34de238874c7192d59a15210faa081