Add support for custom WindowSizeClass values.
Add support for custom WindowSizeClass values.
Open the constructor so devs can create their own WSC values.
Add a function to select the best match from a Set<WindowSizeClass>.
Add utility methods to see if a WindowSizeClass is at least a given
size.

Relnote: "
Update the WindowSizeClass API to support adding new breakpoint values
in the future. Instead of having absolute bounds we use the lower bounds
and recommend devs use lower bound checks when processing a
WindowSizeClass.

The existing WindowWidthSizeClass and WindowHeightSizeClass will be
deprecated as they will not be developed further.
"

Bug: 307793953
Test: ./gradlew window:window-core:test

Change-Id: I014cef212823dc5dfa8bb430fdbe0b505f0b81ff