Fix main thread usage in RoomTrackingLiveData
In aosp/3205550 Room's LiveData implementation was converted to Coroutines but an error was made in that invalidated() was being called from a background thread where as it should be called from the main thread. This CL fixes the issue brining back the threading consistency before the refactor.

Also, improved the compute function usage by using a sealed class instead of null-checking which compute function param was provided.

Bug: 361216129
Test: Existing
Change-Id: I41b09f2d87a03ceeca635341dbb130cab5708d88