Add SupportingPane destination interface
Allows NavDestination subclasses to indicate that
they want to be considered a supporting pane - a
destination that is displayed alongside another
destination.

This allows multiple destinations to simultaneously
be resumed (or, if under a dialog, started).
Essentially, a SupportingPane destination will share
the same lifecycle state as the destination underneath
it.

This is transitive - multiple SupportingPane
destinations can all 'share' their lifecycle.

There is a special exception for cases were a
NavDestination implements both the FloatingWindow
and SupportingPane interfaces, where those kind of
destinations can only 'share' their lifecycle with
other FloatingWindow destinations.

Unlike FloatingWindow destinations, a SupportingPane
destination can be in any location on the back stack.

Relnote: "Custom Navigators can now mark their destinations
as implementing the `SupportingPane` interface which
indicates to the `NavController` that these destinations
will be shown alongside other destinations. By using this
interface, multiple destinations can be `RESUMED` at the
same time, for instance."
Test: newly added tests
BUG: 294612000

Change-Id: Id5559bdaf3e31fb6d6fb66be0b451627c075d342