Fix type names for inner types
For inner types with generic outer types like `A<Foo>.B<Bar>` we generate type name `A.B<Bar, Foo>` or sometimes throw IndexOutOfBoundsException when getting type names. This CL is to fix that.

This is partly due to that `KSType.arguments` on inner types contain type arguments applied to outer types, e.g. `arguments` for `A<Foo>.B<Bar>` is `[Bar, Foo]`.

Bug: 362923743, https://github.com/google/dagger/issues/4409
Test: XTypeElementTest.innerTypeNames
Change-Id: I8883e2ed47d6f2f7c84949a5ff6a775350814c5f