Add mediator overlay flow to manual test app
Adds an additional mediation option that allows the mediator to
overlay the mediatee's view when both parties are in the SDK
runtime. When this is set, 90% of the SandboxedSdkView displayed
in the app will be mediatee content, with the other 10% being
the mediator's overlay.

Also cleaned up some mediation option logic which was unnecessary.

Test: manual test app
Bug: 362467889
Change-Id: Iee5d5863c05f55aed819d3375d2aeed3815d028e