Migrate :compose:animation:animation-core androidUnitTest to commonTest
Not all tests were migrated:
SpringEstimationTest uses Parametrized test runner and there is no alternative for it in commonTest.
One of EasingTest was not migrated because it relies on Float.ulp extensions which is available only for kotlin/jvm.

This change will help Compose Multiplatform to run these tests too.

Test: N/A (no new features or implementation changes)
Change-Id: Id825de2631081b504fa2848f0841babb9ffc377e