Fix a regression when using group optimization
There are cases where the group optimization could
cause the slot table to thrash, moving the slot gap up
and down, for every iteration of a loop, for example,
causing which could cause a signficiant regression
(see the benchmark added previously).

This reduces the thrashing by only adding to the
parent group slots when the parent group closes
instead of whenever the a slot is added to the
group.

Fixes: 362291064
Test: new tests, ./gradlew :compose:r:r:test
Change-Id: I6a592ab2fa77a7964b72335a3076a8faf22c5b3a