Support progress overflow for CircularProgressIndicator and SegmentedCircularProgressIndicator.
Allow overflow  progress values larger than 1.0 (previously progress was
coerced to 0..1), and truncate the value to 0.0..1.0 range with the
modulo operation. For example, 1.2 will be truncated to 0.2 and 2.0 will
be truncated to 1.0.

Screenshot: https://screenshot.googleplex.com/628zgtmB4Tf7koG.png
Relnote: "Support progress overflow (>100% progress) for  CircularProgressIndicator and
SegmentedCircularProgressIndicator."
Bug: 363944011
Test: Added unit tests, screenshot tests, and updated demos.

Change-Id: Iaaa3d4da94307062cd92916be09130bae63afe6a