Adding UUID support in JVM Desktop.
This change simply entails moving the UUIDUtil class to the jvmAndroid source set to make the conversion functions available to both Android and JVM sources. UUIDTest.kt has been added to the JVM source set of the Room KMP test app.

The previous fix to b/362486575 in the TypeAdapterStore has been reverted to enable UUID and ByteBuffer support in JVM Desktop.

Bug: 362994709
Test: UUIDTest.kt
Change-Id: I2330129032ad4351add358fa9799e3a028a21fad