When there are complication overrides, make updates synchronous
This patch makes complication updates synchronous while there are
complication overrides. This fixes the problem reported in the
bug where when editing, the icon from the previous selection was
getting rendered. This happened because complication updates were
asynch (that's usually what you want for interactive rendering but
system editing is a special case because we want the update to finish
before the screen shot is taken).

Test: New unit test, plus verified the fix addresses the problem from the bug.
Bug: 357556142
Change-Id: Ib5da29d4ba34998644e3f19a8181044de9c096b9