Fix ConcurrentModificationException in InvalidationTracker
In commit 01f2d472621978bb38977fb20b63c1d1b6c0995c the WeakObserver was updated to not launch when it needs to be self-removed, which means it can be that during invalidation it self-removes immediately causing a ConcurrentModificationException due to the `observerMap` being modified while also being iterated over. This CL fixes the issue by always making copies of the observer map keys or values before iterating over it.

Bug: 365786165
Test: InvalidationTrackerTest.selfRemovingObserver
Change-Id: Ife81fd6ae49c0fbb1cb02d7629ad2f669a5655ee