Fix a desktop backhandler behavior (#1860)
Change an `Esc` button interception from `onPreviewKeyEvent` to
`onKeyEvent` in the desktop BackGestureDispatcher implementation. It
fixes problems when user's code handles the Esc-button.
Fix a problem when an onBack callback is updated but listener is not.

Fixes: https://youtrack.jetbrains.com/issue/CMP-7569

## Release Notes
### Fixes - Multiple Platforms
- _(prerelease fix)_ Change an `Esc` button interception from
`onPreviewKeyEvent` to `onKeyEvent` in the desktop
`BackGestureDispatcher` implementation. It fixes problems when user's
code handles the `Esc`-button.
- _(prerelease fix)_ Fix a problem when an `onBack` callback is updated
but listener is not.