Fix POM when there is dependency on an artifact with redirection (#1924)
Fixes
https://youtrack.jetbrains.com/issue/CMP-7761/1.8.0dev2186-doesnt-work-with-JPS-and-Maven

The issue was that `pom` points to a root artifact, not to `-desktop`
artifact. :
```
    <dependency>
      <groupId>org.jetbrains.androidx.lifecycle</groupId>
      <artifactId>lifecycle-runtime</artifactId>
      <version>2.8.4</version>
      <scope>runtime</scope>
    </dependency>
```

It is broken because lifecycle-runtime has a redirection, and KGP can't
handle it properly.

We already had the fix for `project` dependencies in
https://github.com/JetBrains/compose-multiplatform-core/pull/1171, but
the issue also exist if we depend on an already published library. In
this PR, we apply the fix for such libraries.

To dermine redirection, we use some heuristic by comparing module names.
If we encounter this
```
[first-level-dependency] org.jetbrains.androidx.lifecycle:lifecycle-runtime:2.8.4 ->
[artifact-with-the-same-name] androidx.lifecycle:lifecycle-runtime:2.8.5 ->
[artifact-with-the-same-name-plus-suffix] androidx.lifecycle:lifecycle-runtime-desktop:2.8.5
```
it means there is a redirection. If it was determined wrongly,
integration tests should fail.

## Testing

### Method 1
1. Perform `./gradlew publishComposeJbToMavenLocal`
2. `mvn install exec:java -Dexec.mainClass="MainKt"
-Dkotlin.version=2.1.0 -Dcompose.version=9999.0.0-SNAPSHOT` in
https://github.com/JetBrains/compose-multiplatform/tree/master/ci/templates/maven-test-project
works

### Method 2
1. Perform `gradlew publishComposeJbToMavenLocal
-Pcompose.platforms=all` for the state before the fix and after the fix
2. See the diff for
[~/.m2/repository/org/jetbrains](https://github.com/JetBrains/compose-multiplatform-core/commit/ab4561c91a3c9f64c5985a5e4539c937e2edf3ad).
Only pinned dependencies should change

## Release Notes
N/A