CMP-7543 Retain AnnotatedString in ClipEntries on Desktop (#2092)
Prior to this, the implementation of `ClipboardUtils.desktop.kt` would
"flatten" `AnnotatedString`s into their plaintext representation. This
means that if you implemented a custom `Clipboard`, it would only
receive the plaintext representation of the copied text in `setEntry`.
With this change, the `AnnotatedString` is actually preserved when
copying, so `Clipboard` can access the rich text and serialize it as it
prefers.

It should not change the observed behaviour when using the default
`Clipboard`, but allows users to create custom ones which can do things
like [JEWEL-815](https://youtrack.jetbrains.com/issue/JEWEL-815).

PS: this removes the `withContext` calls from
`ClipboardUtils.desktop.kt` since they optimise for rare cases
(extremely large content in the clipboard, probably O(100MB) or more)
which are going to clog on the AWT side or during composition anyway. It
can cause a small perf degradation in the overwhelmingly more common
scenarios of tiny-to-large clipboard contents due to the dispatching of
the coroutine to the IO dispatcher. This was discussed with someone from
the Compose Text team.

## Testing
Ran tests, smoke tested by hand. 

## Release Notes
### Features - Desktop
- Full `AnnotatedString` is available as a data flavor in `ClipEntry`,
instead of only its text.

---------

Co-authored-by: Igor Demin <igordmn@users.noreply.github.com>