Use touch event position for the initial positioning of backing HTML input (#2260)
By the time `WebTextInputService.startInput` is called, the real/correct
position for backing HTML input might be not known yet. Therefore
requesting the focus (and thus showing the keyboard) would make the
keyboard overlap the TextField.

We know the real position later, but `input.focus()` request is ignored
unless it's requested in a handler of some user interaction (e.g. touch
event). So we check if startInput is within a touch event handler and
use its x,y as an initial position for the backing HTML input.

Fixes https://youtrack.jetbrains.com/issue/CMP-8611


https://github.com/user-attachments/assets/717d22bc-b8e8-4a20-a8b9-16a7bbd1ad31




## Testing
This should be tested by QA

## Release Notes
### Fixes - Web
- Fix software keyboard behaviour for Compose Text Fields in iOS Safari