CfW Fix Composite text input (#2256)
This PR introduces a different approach to handling the key and input
events.
Due to the differences in the browser behavoiur (events order, key
values, timestamps), we collect the events and process them together in
window.reqiestAnimationFrame callback. It allows to have more info at
one point, which simplifies the processing - the decision making is less
scattered through the code.

Fixes https://youtrack.jetbrains.com/issue/CMP-8416

## Testing
- Added new tests

This should be tested by QA

## Release Notes
### Fixes - Web
- Fixed the bugs with composite text input