Increase CanvasBasedWindow deprecation level to DeprecationLevel.ERROR (#2280)
This PR deprecates CanvasBasedWindow on web platforms, making an attempt
to use it a compilation error


## Testing
`./gradlew testWeb` plus manual

## Release Notes
###  Migration Notes  - Web
- `CanvasBasedWindow` is deprecated, use `ComposeViewport` instead.
Unlike `CanvasBasedWindow`, which expect as an input param the id to the
`HTMLCanvasElement` that will be used for rendering, `ComposeViewport`
one passes `parentContainer` (and corresponding HTML Canvas element will
be created automatically). By default such container is `document.body`.

---------

Co-authored-by: Oleksandr Karpovich <a.n.karpovich@gmail.com>