[web] Introduce more complex drag and drop example (#2515)
Since automated tests for DnD in web is something we don't have, it's
important to have some playground demo for drag-n-drop. We already have
one, the goal of this particular PR is to make it bit more
shiny/complicated.

## Testing
Launch demo and check the Drag and Drop example

## Release Notes
N/A