Update accessibility label for merged nodes (#2539)
Implement logic to merge the content descriptions for accessibility
elements that mimics the behaviour of the TalkBack app. Use a comma
instead of a newline symbol to join content descriptions.
Fix an issue where accessibility elements inside other accessibility
elements may not be accessible.

Fixes
https://youtrack.jetbrains.com/issue/CMP-9124/iOS-A11y-mergeDescendants-custom-contentDescription

## Release Notes
### Fixes - iOS
- Fix the speaking of text in merged accessibility nodes.
- Fix an issue where accessibility elements inside other accessibility
elements may not be accessible.