Fix runParametrizedTest implementation (#2640)
The existing implementation violated the requirement that a test method
must return the TestResult -
https://kotlinlang.org/api/kotlinx.coroutines/kotlinx-coroutines-test/kotlinx.coroutines.test/run-test.html

> On JS, this function creates a Promise that executes the test body
with the delay-skipping behavior.
The platform difference entails that, in order to use this function
correctly in common code, one must always immediately return the
produced
[TestResult](https://kotlinlang.org/api/kotlinx.coroutines/kotlinx-coroutines-test/kotlinx.coroutines.test/-test-result/index.html)
from the test method, without doing anything else afterwards.

## Testing
Existing tests

## Release Notes
N/A